Add files via upload
The objective of this analysis is Soft-Drink  Demand Quantity prediction which aims to identify the relationship between the soft drink quantity in Greece Markets supplies to the vendors . In addition, predict the city , brand , containers.. quantity is the most valuable variable to predict  . The project involves predictions using two  regression techniques Decision tree and Random forest. Demand Quantity on a dataset has been done by using the above mention techniques to find out the best among the two.  The motive of this project is to help  The built model has evaluated with test data and verified that the model predicted the Quantity with minimum error rate.